Refactor/debug FormFactorAnalysis. Change how tmax inferred.
With these changes, FormFactorAnalysis now seems to produce working
fits. This refactoring required many big changes in the analysis
module and a variety of smaller ones elsewhere. Most of the plotting
routines in this module have now been refactored and checked. Lots
of formatting changes are still necessary for PEP8 / pylint.

In the correlator module, I corrected an edge case so that tmax is
now inferred more sensibly when noise_threshy is None.