Profile and optimize correct_covariance, add TODO list to README
The function correct_covariance was running very slowly. We had
thought this might be an issue with some of the shrinkage routines.
Following some profiling, the true culprit turned out to be
np.einsum, which was running surprisingly slowly. Switching to
dedicated matrix multiplication routines gave an ~100x speed up.