Add logging, remove print statements, and fix pylint formatting
Support for logging instead of printing is now included. Basically,
an application (script, notebook, etc...) which uses these modules
should invoke something like the following:
import logging
import logging.config
logging.config.fileConfig(
    'logging_config.ini',
    disable_existing_loggers=False
)
The user is responsible for making sure that their initialization
file is on the path. The included 'logging_config.ini' file is just
one possible example. This usage decouples the library from the
formatting and handling of the logs, which is considered a best
practice.