Generalize initialization of TwoPoint and failure mode for FastFit
The previous initialization of TwoPoint lacked any possibility to
tweak the times associated with the correlator. This is an OK
default, but it was too rigid for noisy correlators, like those
from incomplete runs. The new version supports kwargs which are
passed to BaseTimes.

When FastFit fails because of "cosh(E) < 1", initialization now
succeeds, and the energy and amplitude are set to None. This
change allows FastFit to be used more robustly with noisy data
(where failures may occur unexpectedly). I added a logger to
FastFit, which issues a warning for "cosh(E) < 1" errors.