Correct initialization of FormFactorDataset
The former implementation did not allow for different correlators
to have different tdata. This was a major bug, since the primary
usage case is to fold two-point correlators about the midpoint.
To keep the initialization simple, nt for the two-point functions
is inferred from the three-point function(s). This will always
hold in practice, since it's only sensible to analyze correlators
from the same ensemble, which has fixed nt. More detailed checks
are not appropriate here.