Add format()
The included format() routine is basically copied from lsqfit,
but with some important differences. First, it no longer throws
an error if it encounters a NaN somewhere in the output table.
Second, it now prints both Bayesian and frequentist summary
statistics at the top, saving the need to compute them "by hand"
when people ask.