Switch FormFactorDataset._tags to public member
ds._tags turns out to be a useful attribute. It was a design mistake
to make this member nominally private by naming it with an underscore.
In this commit, I've changed the actual attribute to be ds.tags. To
maintain temporary backward compatibility, a new property mimicks
the old functionality but issues a deprecation warning. In the future,
support will be removed.