Simplify definition of C3bar
The quantity C3bar is an "average over adjacent time slices."
In the literature, for a given 3pt function C(t, T)one encounters
averaging over the variables t (the Euclidean time of the "current")
and T (the Euclidean time of the "sink"). In typical datasets, t is
varied, while T is held fixed at a few carefully chosen values. In
practice, this means that averaging in t is trivial and unambiguous,
while averaging in T requires some choices (e.g., "How should one
handle combining non-adjacent values for T?"). For direct fits to the
ratio Rbar, it is important that averaging be done consistently for all
T values. This treatment makes sure that all sinks are treated
consistently and on equal footing.