Update inputs, add support for inuts in MeV
For constructing final results, it's necessary (or at least useful) use
a fixed set of inputs when evaluating the form factors. This treatment
allows for consistent tracking of errors from quantities like fpi, w0,
and so on. The upshot is that convenient to be able to specify inputs
directly in physical units and them convert them to intermediate units
(like w0) internally.

This commit also tweaks the set of input parameters which are required
by ChiralExpansionParameters.