Refactor Zfitter, fix bug in tplus
Add support for BESIII’s choice for t0.
Add support for q2(z).
Correct bug in definition of tplus. Recall that fplus is the
location of the 2-particle cut, not necessarily the sum of
mother and daughter masses. This distinction is relevant
for Ds2K, where the cut begins and the D-pi threshold.